tidying and consistency changes
moved calling of galaxy display to gameController, which is created and invoked by the Loaded() function in GalaxyGenGUIController. changes will need to be made so that the player starts inside a system and not on the galaxy map. galaxy generation will be modified to populate the galaxy with factions + space stations. the player will start at a non hostile space station and be able to access the galaxy map through the game's GUI. other general tidy up and refactoring of code to improve readability and consistency was performed as well