Factions started
Factions are being generated in GalaxyGen in the galaxy generator. Desire Values for each type of faction are pre calculated as a means to decide what system are optimal for each faction type. Each type with a number of system configurations have been evaluated through unit tests. if adjustments are made to the desire value used for each body type then the expected values will need to be updated