Flight
groundwork for space flight has been added. ships retain their momentum and thrust is only necessary to change the vector of travel. the flight model takes into account the lorentz factor for limiting acceleration towards a maximum speed. to avoid loss of control at high speeds this is only taken into account when the thrust vector is +-90 degrees from the ship's current velocity. this allows for ships to slow down at their maximum rate of acceleration and incentivizes using the main thrusters as the movement differential and maneouvring thrusters for smaller course corrections. the classes for ships and their components are designed to be modular so thrusters(and later other components) may be added or removed for tweaking performance. the force thrusters can put out will need to be tweaked and possibly changed so that they spin up over time rather than providing their maximum force at all times