turret rotation
implemented ship turret rotation for the player. needs to be changed to allow for AI to aim at specific points instead of using the mouse position. as the ship controller will be handling aiming through each weapon controller the input vector will need to be passed as a parameter to the turnWeapon function