Ship Outfitting
Ship Outfitting GUI pulls ship components from OutfittingService which will eventually be attached to space stations. Currently it is instatiated with all currently created ShipComponent types. Later different stations will have different items available. The outfitting GUi is currently initialised with 1 list of all the components. This will need to be modified to display only items according to the buttons on the components panel. Attaching ship hulls for modification with the components in the list is the next step after this