Outfitting Ship Display
Now displaying the model of the current ship in the outfitting GUI. will need to ensure that the ships are centered to the camera when the camera is not at the origin of the scene. need to add colliders for hovering over different parts of the ship and selecting them for changing parts. parts when selected will need to be instantiated and displayed using their models on the ship itself in addition to previously specified work to set components to their specific menus