Dynamic Ship Creation
Added ShipCreator class to handle building and loadout randomisation of ships within the GameController. Different permutations of slots can be used and different amounts of weapons/thrusters can be added onto the ships. Factions have lists of components that they are allowed to use with weights to guide the randomisation more or less heavily to specific components in the allowed lists. More hull types/weapons are needed for true variety. Internal components need to be added to the CreateShip function in the ShipCreator