Faction Desire Changes & Space Stations
Changed how the desire values are calculated for SolarSystems and Celestial Bodies - reworked the tests to adjust for these changes. Now the main Space Station for a system is set dependant on the most desirable CelestialBody within the solarSystem. Starting system is selected using the homeworld of a random faction from a list of eligible faction types