Station GUI & Outfitting
Made the station GUI work by populating the menu with the station services assigned to it. OutfittingGUi is now instantiated through it's controller and the controller is no longer attached to the GUI prefab. The outfitting service is populated with faction specfic components. Other components need to be added such as cargo bays, shields etc. Adding filtering for these lists will be an important feature for improving QoL and UX. OutfittingGUI needs to be adjusted so that it can be closed and route back to the station GUI. The rest of the station services should be implemented alongside these improvements