Tidying execution, Script Parents & Outfitting Closing
Removed creation of prefabHandlers as and when needed and moved it to a static variable in the GameController to reduce unecessary duplication. Tied certain scripts to the gameobjects that are using them to reduce duplication on opening and closing GUIs and other repeated functions. Made outfitting able to be closed using a home button that takes the user back to the StationGUI