Commodities & Trading
Implemented various commodities for trading. Different factions have affinities for different types of products and are produced per solar system depending on the planet types within. TradeGUI allows adding commodities to the cargo bays in the current ship and updates the tradeService accordingly. Selling cargo still needs to be implemented